% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sync_status.R
\name{sync_status}
\alias{sync_status}
\title{Check Project Status and Sync}
\usage{
sync_status()
}
\description{
Opens an interactive Shiny Gadget allowing users to select files to publish
via Workflowr or commit/push via Git. Automatically checks for unpulled changes
and verifies final synchronization status. Features a 2-step review process.
}
